Update AdminDashboard
Implementation of a confirmation dialog before deleting a user or department

+++ ethics-g4-app/client/src/pages/AdminDashboard.jsx
@@ -45,21 +45,26 @@ const AdminDashboard = () => {
   );
 
   const deleteUser = (userId) => {
-    fetch(`${import.meta.env.VITE_SERVER_URL}/api/users/${userId}/delete/`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error("Failed to delete user");
-        }
-        // Refresh the users list or handle UI update
-        setFetchTrigger((prev) => prev + 1);
+    // Display confirmation dialog before deleting user
+    const confirmDelete = window.confirm("Are you sure you want to delete this user?");
+    if (confirmDelete) {
+      fetch(`${import.meta.env.VITE_SERVER_URL}/api/users/${userId}/delete/`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
       })
-      .catch((error) => console.error("Error deleting user:", error));
+        .then((response) => {
+          if (!response.ok) {
+            throw new Error("Failed to delete user");
+          }
+          // Refresh the users list or handle UI update
+          setFetchTrigger((prev) => prev + 1);
+        })
+        .catch((error) => console.error("Error deleting user:", error));
+    }
   };
+  
 
   // Function to update user role
   const updateUserRole = (userId, newRole, departmentCode) => {
@@ -115,19 +120,23 @@ const AdminDashboard = () => {
   
 
   const deleteDepartment = async (departmentId) => {
-    try {
-      const response = await fetch(
-        `${import.meta.env.VITE_SERVER_URL}/api/departments/${departmentId}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      );
-      setFetchTrigger((prev) => prev + 1);
-    } catch (error) {
-      console.error("Error deleting department:", error.message);
+    // Display confirmation dialog before deleting department
+    const confirmDelete = window.confirm("Are you sure you want to delete this department?");
+    if (confirmDelete) {
+      try {
+        const response = await fetch(
+          `${import.meta.env.VITE_SERVER_URL}/api/departments/${departmentId}`,
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }
+        );
+        setFetchTrigger((prev) => prev + 1);
+      } catch (error) {
+        console.error("Error deleting department:", error.message);
+      }
     }
   };
